+++ src/main/java/name/divinityunbound/DivinityUnboundClient.java
@@ -46,6 +46,8 @@ public class DivinityUnboundClient implements ClientModInitializer {
         HandledScreens.register(ModScreenHandlers.FLUID_TRASHCAN_SCREEN_HANDLER, FluidTrashcanScreen::new);
         HandledScreens.register(ModScreenHandlers.ENERGY_TRASHCAN_SCREEN_HANDLER, EnergyTrashcanScreen::new);
         HandledScreens.register(ModScreenHandlers.HALLOWED_FLUID_TANK_SCREEN_HANDLER, HallowedFluidTankScreen::new);
+        //HandledScreens.register(ModScreenHandlers.ITEM_SINGULARITY_STORAGE_SCREEN_HANDLER, ItemSingularityStorageScreen::new);
+        HandledScreens.register(ModScreenHandlers.KNOWLEDGE_EXTRACTOR_SCREEN_HANDLER, KnowledgeExtractorScreen::new);
         HandledScreens.register(ModScreenHandlers.SPACE_TIME_FURNACE_SCREEN_HANDLER, SpaceTimeFurnaceScreen::new);
 
         BlockEntityRendererFactories.register(ModBlockEntities.SPACE_SIPHON_BLOCK_ENTITY, SpaceSiphonBlockRenderer::new);