+++ src/main/java/name/divinityunbound/datagen/ModLootTableProvider.java
@@ -66,6 +66,8 @@ public class ModLootTableProvider extends FabricBlockLootTableProvider {
         addDrop(ModBlocks.FLUID_TRASHCAN);
         addDrop(ModBlocks.ENERGY_TRASHCAN);
         addDrop(ModBlocks.HALLOWED_FLUID_TANK);
+        addDrop(ModBlocks.ITEM_SINGULARITY_STORAGE);
+        addDrop(ModBlocks.KNOWLEDGE_EXTRACTOR);
 
         addDrop(ModBlocks.SPACE_TIME_FURNACE);
     }