+++ src/main/java/name/divinityunbound/item/ModItemGroups.java
@@ -151,6 +151,7 @@ public class ModItemGroups {
                         entries.add(ModBlocks.FLUID_TRASHCAN);
                         entries.add(ModBlocks.ENERGY_TRASHCAN);
                         entries.add(ModBlocks.HALLOWED_FLUID_TANK);
+                        //entries.add(ModBlocks.ITEM_SINGULARITY_STORAGE);
 
                         entries.add(ModBlocks.SPACE_TIME_FURNACE);