+++ src/main/java/name/divinityunbound/screen/KnowledgeExtractorScreenHandler.java
@@ -1,0 +1,62 @@
+package name.divinityunbound.screen;
+
+import name.divinityunbound.block.entity.GenerationStationBlockEntity;
+import name.divinityunbound.block.entity.KnowledgeExtractorBlockEntity;
+import net.minecraft.block.entity.BlockEntity;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.entity.player.PlayerInventory;
+import net.minecraft.inventory.Inventory;
+import net.minecraft.item.ItemStack;
+import net.minecraft.network.PacketByteBuf;
+import net.minecraft.screen.ArrayPropertyDelegate;
+import net.minecraft.screen.PropertyDelegate;
+import net.minecraft.screen.ScreenHandler;
+import net.minecraft.screen.slot.Slot;
+
+public class KnowledgeExtractorScreenHandler extends ScreenHandler {
+    private final PropertyDelegate propertyDelegate;
+    public final KnowledgeExtractorBlockEntity blockEntity;
+
+    public KnowledgeExtractorScreenHandler(int syncId, PlayerInventory inventory, PacketByteBuf buf) {
+        this(syncId, inventory, inventory.player.getWorld().getBlockEntity(buf.readBlockPos()),
+                new ArrayPropertyDelegate(2));
+    }
+
+    public KnowledgeExtractorScreenHandler(int syncId, PlayerInventory playerInventory, BlockEntity blockEntity, PropertyDelegate arrayPropertyDelegate) {
+        super(ModScreenHandlers.KNOWLEDGE_EXTRACTOR_SCREEN_HANDLER, syncId);
+        this.propertyDelegate = arrayPropertyDelegate;
+        this.blockEntity = (KnowledgeExtractorBlockEntity) blockEntity;
+
+        addPlayerInventory(playerInventory);
+        addPlayerHotbar(playerInventory);
+
+        addProperties(arrayPropertyDelegate);
+    }
+
+    public int getStoredExperienceLevel() { return this.propertyDelegate.get(0); }
+    public int getStoredTotalExperience() { return this.propertyDelegate.get(1); }
+
+    @Override
+    public ItemStack quickMove(PlayerEntity player, int invSlot) {
+        return ItemStack.EMPTY;
+    }
+
+    @Override
+    public boolean canUse(PlayerEntity player) {
+        return true;
+    }
+
+    private void addPlayerInventory(PlayerInventory playerInventory) {
+        for (int i = 0; i < 3; ++i) {
+            for (int l = 0; l < 9; ++l) {
+                this.addSlot(new Slot(playerInventory, l + i * 9 + 9, 8 + l * 18, 84 + i * 18));
+            }
+        }
+    }
+
+    private void addPlayerHotbar(PlayerInventory playerInventory) {
+        for (int i = 0; i < 9; ++i) {
+            this.addSlot(new Slot(playerInventory, i, 8 + i * 18, 142));
+        }
+    }
+}