+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ---
 ## v0.1.0
-- Updates to several block GUIs to make specific inputs and outputs more clear 
 - Added Hallowed Fluid Tank
+- Added Knowledge Extractor (can store XP)
+- Updates to several block GUIs to make specific inputs and outputs more clear
 - Fixed import/export fluid bug with Space Time Evaporator and Amalgamator
+- Bug Fixes (IOCard can now only save coords of Wormhole Transporter blocks)
 
 ---
 ## v0.0.7